requirements.txt
absl-py==2.1.0
accelerate==0.29.3
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
av==12.0.0
beautifulsoup4==4.12.3
blessed==1.20.0
click==8.1.7
cmake==3.30.2
contourpy==1.2.0
cycler==0.12.1
easydict==1.13
einops==0.7.0
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
ftfy==6.2.0
gast==0.6.0
gdown==5.2.0
glcontext==2.5.0
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.62.1
h5py==3.11.0
huggingface-hub==0.22.2
imageio==2.34.0
imageio-ffmpeg==0.4.9
joblib==1.4.2
keras==3.5.0
kiui==0.2.11
kiwisolver==1.4.5
lazy-loader==0.4
libclang==18.1.1
lit==18.1.8
lpips==0.1.4
markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.4.0
moderngl==5.10.0
namex==0.0.8
nvidia-ml-py==12.560.30
objprint==0.2.3
omegaconf==2.3.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.12.1
packaging==24.0
pandas==2.2.2
plyfile==1.0.3
prettytable==3.10.0
protobuf==4.25.0
psutil==5.9.8
pyglm==2.7.1
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
regex==2023.12.25
rich==13.7.1
safetensors==0.4.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
torch-tb-profiler==0.4.3
tqdm==4.66.2
trimesh==4.3.1
triton==2.0.0
tzdata==2024.1
varname==0.13.3
wcwidth==0.2.13
werkzeug==3.0.1
wrapt==1.16.0
xformers==0.0.22
